+++ src/main/java/me/hp20/giz5/options/HudOptions.java
@@ -1,15 +1,15 @@
 package me.hp20.giz5.options;
 
 public class HudOptions {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     public Position position;
 
-    public int getX() {
+    public float getX() {
         return x + position.getX();
     }
 
-    public int getY() {
+    public float getY() {
         return y + position.getY();
     }
 }
